pyproject.toml
[tool.poetry]
name = "opsi-deploy-client-agent"
version = "4.2.0.22"
description = "tool to deploy the opsi-client-agent on linux, windows and macOS clients"
homepage = "https://www.opsi.org"
license = "AGPL-3.0"
authors = ["uib GmbH <info@uib.de>"]
maintainers = ["uib GmbH <info@uib.de>"]
packages = [
	{ include = "opsideployclientagent" }
]

[[tool.poetry.source]]
name = "uib"
url = "https://pypi.uib.gmbh/simple/"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
python-opsi = "^4.2.0.274"
python-opsi-common = "^4.2.0.275"
paramiko = "^2.9"
impacket = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pyinstaller = "^4.9"
pylint = "^2.12"
tan = "^22.2"
flake8 = "^4.0"
mypy = "^0.931"

[tool.poetry.scripts]
opsi-deploy-client-agent = "opsideployclientagent.__main__:main"

[tool.black]
line-length = 140
target-version = ['py39', 'py310']
use-tabs = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
